Questions/BinaryTree/101-Symmetric-Tree/symmetric.go
package binarytree

import "fmt"

type TreeNode struct {
	Val   int
	Left  *TreeNode
	Right *TreeNode
}

func isSymmetric(root *TreeNode) bool {
	if root == nil {
		return true
	}

	return checkSymmetry(root, root)
}

func checkSymmetry(node1 *TreeNode, node2 *TreeNode) bool {
	if node1 == nil && node2 == nil {
		return true
	} else if node1 == nil || node2 == nil {
		return false
	}

	return node1.Val == node2.Val && checkSymmetry(node1.Left, node2.Right) && checkSymmetry(node1.Right, node2.Left)
}

func main() {
	root := &TreeNode{Val: 1}
	root.Left = &TreeNode{Val: 2}
	root.Right = &TreeNode{Val: 2}
	root.Right.Left = &TreeNode{Val: 3}
	root.Right.Right = &TreeNode{Val: 4}
	root.Left.Left = &TreeNode{Val: 4}
	root.Left.Right = &TreeNode{Val: 3}

	fmt.Printf("Ans: %v\n", isSymmetric(root))
}
